requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.1
attrs==25.3.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
boto3==1.37.33
botocore==1.37.33
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
cryptography==44.0.1
deep-translator==1.11.4
defusedxml==0.7.1
edge-tts==7.0.0
fastjsonschema==2.21.1
firebase-admin==6.6.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
Mako==1.3.9
MarkupSafe==3.0.2
mistune==3.1.3
msgpack==1.1.0
multidict==6.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pypiwin32==223
pyspark==3.5.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pywin32==308
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.38
srt==3.5.3
tabulate==0.9.0
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
yarl==1.18.3
